Added Engineering Journal and Notes
