+++ README.md
@@ -1,4 +1,4 @@
-# zillowish-clifton
+# Zillowish-Facts-Features
 clifton's portion
 
 Related Projects:
@@ -23,3 +23,21 @@ To Seed the Databse:
 `npm run db-seed`
 
 This will create 100 seeds into the mysql database.
+
+##SDC Engineering Journal and Notes
+
+###Database Benchmarking
+
+Which two DBMS did you test?
+SQL: POSTGRE
+noSQL: MONGODB
+
+This is what happened when I tried to load 10M records without any optimizations.
+FATAL ERROR: Ineffective mark-compacts near heap limit Allocation failed - JavaScript heap out of memory
+###Performance notes.
+[TODO]
+###Stress Testing
+[TODO]
+###Other Notes
+[TODO]
+
