+++ database/seedDatabase.js
@@ -1,18 +1,22 @@
 const mock = require('./createMockData');
 const db = require('./index');
 
+db.createDatabase(function () {
 
-db.loadFeatures((err)=> {
-  if (err) {
-    throw err;
-  }
+  db.loadFeatures((err)=> {
+    if (err) {
+      throw err;
+    }
+  });
+
+  db.loadInterior((err)=> {
+    if (err) {
+      throw err;
+    }
+  });
+
 });
 
-db.loadInterior((err)=> {
-  if (err) {
-    throw err;
-  }
-});
