+++ server/server.js
@@ -4,12 +4,11 @@ const app = express();
 const port = 3003;
 const bodyParser = require('body-parser');
 const data = require('../database/index');
-const features = require('../database/createMockData');
 
-
+data.createDatabase(()=>{});
 app.use(bodyParser.urlencoded({ extended: false }));
 app.use(bodyParser.json());
-app.use(cors({origin: 'http://localhost:3003'}));
+app.use(cors());
 app.use(express.static(__dirname+'/../client/dist', {maxAge: 5000}));
 
 app.get('/house/:id', (req, res) => {
